fix(Uikit-playground): fix contextual bar
