+++ packages/uikit-playground/src/Components/FlowContainer/UIKitWrapper/UIKitWrapper.scss
@@ -1,0 +1,7 @@
+.rc-scrollbars-view{
+    position: relative !important;
+    overflow: hidden !important;
+    width: 100% !important;
+    z-index: 1000 !important;
+    margin: 0 !important;
+}
