+++ packages/uikit-playground/src/Components/FlowContainer/UIKitWrapper/index.ts
@@ -1,0 +1,1 @@
+export {default} from './UIKitWrapper';
